Add some prototypes for an obfuscated store
The basic idea is that files and directories will be given
pure random names.  An index will define mappings from name
to random name.  The index will be encrypted against the
keys that have access permission for the directory that
the node resides in.  (Note that a subdirectory is considered
to have the access of the keys within it -- this way the
users of that node can see its name.)

TODO: Also consider integration with the realms concept.  Per-
directory realms is probably sufficient, but per-file realm
controls might be helpful.